refactor: rewrite footnotes
